+++ src/sound/ogg_sound_file.cpp
@@ -87,7 +87,7 @@ OggSoundFile::OggSoundFile(const Pathname& filename) :
 
   vorbis_info* vi = ov_info(&m_vorbis_file, -1);
   m_channels = vi->channels;
-  m_rate = vi->rate;
+  m_rate = static_cast<int>(vi->rate);
   m_bits_per_sample = 16;
   m_size = static_cast<size_t> (ov_pcm_total(&m_vorbis_file, -1) * 2);
 }
